Add support for AMI BlindTransfer Action
This pull request adds support for BlindTransfer, an AMI action that allows the connected channel to be transferred to another context/extension within the dial plan. To use this, you must call it like so:

```javascript
new nami.Action.BlindTransfer('channel', 'context', 'extension');
```

The channel that is referred to in the parameters is not the channel to be transferred, rather the channel that wants to do the transferring. For example, an Agent is connected to a caller, and he wants to transfer that caller to another context/extension, you have to pass in the Agent's channel name.