Updated API for QueuePause and QueueUnpaused
I have updated the API for QueuePause and QueueUnpause to allow parameter injection rather than messing with the internal instance properties. This is safer, as someone could forget to instantiate the Action, which would result in the context referred to inside the action being the global scope rather than the instance. 

I also fixed a bug where QueueUnpause() would send 'QueueUnpause' as an action, rather than 'QueuePause' with paused = false.